requirements.txt
aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
jinja2>=2.10.1,<3.0.0
